Removed internet connection check
An internet access check was still in the code from when it was originally just a file on a computer (before it was a website) and had to be updated by running a file that came with the program. The website would check for internet access and display a no internet sign if there wasn't an internet connection to alert the user that even if they had the auto-check updates box checked, they wouldn't receive a notification.